fix with latest MinGW & Qt5.15.3

+++ src/fqterm/fqterm_window.h
@@ -113,7 +113,7 @@ private:
 #endif //HAVE_PYTHON
 
 public:
-  virtual long windowID() {return long(this);}
+  virtual long long windowID() {return (long long)this;}
   //end FQTermScriptEventListener
 
 signals:
